v0.4.0(minor): RTX 5090 训练加速 + 注释审计修复
训练加速:
- seed.py: cudnn.benchmark=True, TF32启用
- train.py: find_unused_parameters=False, torch.compile, persistent_workers
- trainer.py: BF16 autocast
- configs: batch_size 4→8, num_workers 8→12

注释修复:
- m3f_detr.py: 192ch→ch[0], backbone_name参数
- test_dataloader.py/matcher.py: 300→num_queries/Nq
- evaluator.py: 删除残留import注释
- depth_encoder.py: normal→gradient术语统一
- simple_fpn.py/dino_head.py: 阶段注释更新

+++ engine/trainer.py
@@ -82,8 +82,8 @@ def train_one_epoch(
 
         optimizer.zero_grad()
 
-        # AMP 前向 + 损失
-        with autocast(enabled=use_amp):
+        # AMP 前向 + 损失（BF16 优于 FP16: 同指数范围无梯度溢出，RTX 5090 原生支持）
+        with autocast(enabled=use_amp, dtype=torch.bfloat16):
             outputs = model(rgb, ir, depth, targets)
             loss_dict = criterion(outputs, targets)
             loss = loss_dict["loss"]
